fix yosys post processing
