+++ siliconcompiler/tools/yosys/syn_asic.py
@@ -1,13 +1,12 @@
 
-from .syn import setup as setup_syn
-from .yosys import setup_asic, prepare_synthesis_libraries, create_abc_synthesis_constraints
+from .yosys import syn_setup, setup_asic, prepare_synthesis_libraries, create_abc_synthesis_constraints, syn_post_process
 
 def setup(chip):
     ''' Helper method for configs specific to ASIC synthesis tasks.
     '''
 
     # Generic synthesis task setup.
-    setup_syn(chip)
+    syn_setup(chip)
 
     # ASIC-specific setup.
     setup_asic(chip)
@@ -20,3 +19,6 @@ def pre_process(chip):
     prepare_synthesis_libraries(chip)
     create_abc_synthesis_constraints(chip)
     return
+
+def post_process(chip):
+    syn_post_process(chip)
