+++ siliconcompiler/tools/yosys/syn_fpga.py
@@ -1,13 +1,12 @@
 
-from .syn import setup as setup_syn
-from .yosys import create_vpr_lib, setup_fpga
+from .yosys import syn_setup, create_vpr_lib, setup_fpga, syn_post_process
 
 def setup(chip):
     ''' Helper method for configs specific to FPGA synthesis tasks.
     '''
 
     # Generic synthesis task setup.
-    setup_syn(chip)
+    syn_setup(chip)
 
     # FPGA-specific setup.
     setup_fpga(chip)
@@ -21,3 +20,7 @@ def pre_process(chip):
     if chip.get('fpga', 'arch'):
         create_vpr_lib(chip)
         return
+
+##################################################
+def post_process(chip):
+    syn_post_process(chip)
